test(weather): Move component tests to feature-specific directory
- Relocated component test files (e.g., app_back_button_test.dart, city_search_button_test.dart, city_search_input_test.dart)
to test/features/weather/components/ for improved modularity.
- Updated relative imports for mocks and providers accordingly.
